add test cases for non-printable ASCII
This aspect of the new spec wasn't properly covered.

cater for \n, \r and \t in string values

Also disallow non-printable chars in tag keys, tag values and
field keys and add test cases for non-printable ASCII
(not properly covered before)

In the [line-protcol proposal](xxx), we said that we'd interpret
`\n`, `\r` and `\t` values in string field values, but that wasn't actually implemented.
Adjust the test corpus now that it is.